requirements.txt
asgiref==3.7.2
Django==3.2.22
django-filter==23.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.5.1
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
